requirements.txt
tensorflow==2.18.0
transformers==4.48.3
torch==2.5.1
torchvision==0.20.1
scikit-learn==1.6.1
numpy==1.26.4
pandas==2.2.2
matplotlib==3.10.0
opencv-python==4.11.0.86
joblib==1.4.2
pillow==11.1.0
tensorflow-hub==0.16.1
cuda-python==12.6.0
streamlit==1.41.1
tensorflow-hub==0.16.1
